Consolidate shared codebase into @tailor-cms/common
