+++ apps/backend/index.js
@@ -26,7 +26,6 @@ const runApp = promisify(app.listen.bind(app));
 database
   .initialize()
   .then(() => logger.info('Database initialized'))
-  .then(() => import('tailor-config-shared'))
   .then(() => contentPluginRegistry.initialize())
   .then(() => runApp(config.port))
   .then(() => {
