+++ apps/backend/package.json
@@ -88,9 +88,8 @@
     "sequelize": "^6.37.4",
     "sequelize-replace-enum-postgres": "^1.6.0",
     "simple-oauth2": "^5.1.0",
-    "sse-event-types": "workspace:*",
     "strip-json-comments": "^5.0.1",
-    "tailor-config-shared": "workspace:*",
+    "@tailor-cms/common": "workspace:*",
     "tailor-seed": "workspace:*",
     "tar-fs": "^3.0.6",
     "tmp": "0.2.1",
