+++ apps/backend/shared/transfer/default/processors.js
@@ -10,7 +10,7 @@ import omit from 'lodash/omit.js';
 import { parse } from 'JSONStream';
 import Promise from 'bluebird';
 import reduce from 'lodash/reduce.js';
-import roleConfig from 'tailor-config-shared/src/role.js';
+import roleConfig from '@tailor-cms/common/src/role.js';
 import { SCHEMAS } from '@tailor-cms/config';
 import zipObject from 'lodash/zipObject.js';
