+++ apps/frontend/components/admin/UserDialog.vue
@@ -81,7 +81,7 @@
 import { object, string } from 'yup';
 import isEmpty from 'lodash/isEmpty';
 import map from 'lodash/map';
-import { role } from 'tailor-config-shared';
+import { role } from '@tailor-cms/common';
 import { TailorDialog } from '@tailor-cms/core-components';
 import { title as titleCase } from 'to-case';
 import { useForm } from 'vee-validate';
