+++ apps/frontend/components/repository/Workflow/Overview/index.vue
@@ -38,7 +38,7 @@
 </template>
 
 <script lang="ts" setup>
-import type { StatusConfig } from 'tailor-config-shared';
+import type { StatusConfig } from '@tailor-cms/common';
 import type { User } from '@tailor-cms/interfaces/user';
 import { UserAvatar } from '@tailor-cms/core-components';
 import { workflow as workflowConfig } from '@tailor-cms/config';
