+++ apps/frontend/package.json
@@ -56,8 +56,7 @@
     "pinia": "^2.2.4",
     "pluralize": "^8.0.0",
     "promise-queue": "^2.2.5",
-    "sse-event-types": "workspace:*",
-    "tailor-config-shared": "workspace:*",
+    "@tailor-cms/common": "workspace:*",
     "to-case": "^2.0.0",
     "url-join": "^5.0.0",
     "v-clipboard": "3.0.0-next.1",
