+++ apps/frontend/pages/repository/[[id]]/root/settings/user-management.vue
@@ -9,7 +9,7 @@
 
 <script lang="ts" setup>
 import map from 'lodash/map';
-import { role } from 'tailor-config-shared';
+import { role } from '@tailor-cms/common';
 import { title as titleCase } from 'to-case';
 
 import AddUserDialog from '@/components/repository/Settings/UserManagement/AddUserDialog.vue';
