+++ apps/frontend/stores/activity.ts
@@ -5,7 +5,7 @@ import {
   InsertLocation,
 } from '@tailor-cms/utils';
 import { schema, workflow as workflowConfig } from '@tailor-cms/config';
-import { Activity as Events } from 'sse-event-types';
+import { Activity as Events } from '@tailor-cms/common/src/sse.js';
 import findIndex from 'lodash/findIndex';
 import Hashids from 'hashids';
 import orderBy from 'lodash/orderBy';
