+++ apps/frontend/stores/content-elements.ts
@@ -1,6 +1,6 @@
 import { calculatePosition } from '@tailor-cms/utils';
 import type { ContentElement } from '@tailor-cms/interfaces/content-element';
-import { ContentElement as Events } from 'sse-event-types';
+import { ContentElement as Events } from '@tailor-cms/common/src/sse.js';
 import flatMap from 'lodash/flatMap';
 import flatten from 'lodash/flatten';
