--- packages/sse-event-types/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "sse-event-types",
-  "type": "module",
-  "main": "index.js"
-}
