+++ packages/tailor-common/src/sse.js
